Create class to clear last*date
Add delete trigger to ContentDocumentLink
Rearrange python code generation and add clear code
